+++ src/test/java/ru/netology/stats/StatServiceTest.java
@@ -17,7 +17,7 @@ public class StatServiceTest {
     @Test
     void statAverageSumSales() {
         StatService service = new StatService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         double expected = 15;
         double actual = service.averageSumSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -26,7 +26,7 @@ public class StatServiceTest {
     @Test
     void statMaxMonth() {
         StatService service = new StatService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 8;
         long actual = service.maxSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -35,7 +35,7 @@ public class StatServiceTest {
     @Test
     void statMinMonth() {
         StatService service = new StatService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 9;
         long actual = service.minSales(sales);
         Assertions.assertEquals(expected, actual);
@@ -44,7 +44,7 @@ public class StatServiceTest {
     @Test
     void statSalesBelowAverage() {
         StatService service = new StatService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         long actual = service.salesBelowAverage(sales);
         Assertions.assertEquals(expected, actual);
@@ -53,7 +53,7 @@ public class StatServiceTest {
     @Test
     void statSalesAboveAverage() {
         StatService service = new StatService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
         int expected = 5;
         long actual = service.salesAboveAverage(sales);
         Assertions.assertEquals(expected, actual);
